+++ esmodules/dorako-ux/radial-condition-hud.js
@@ -89,20 +89,27 @@ function sizeToIconScale(size) {
 function drawBG(effectIcon, background, gridScale) {
   const r = effectIcon.width / 2;
   const isDorakoUiActive = game.modules.get("pf2e-dorako-ui")?.active;
-  const theme = isDorakoUiActive ? game.settings.get("pf2e-dorako-ui", "theme.app-theme") : false;
-  if (theme === "foundry2-theme") {
+  const appTheme = isDorakoUiActive ? game.settings.get("pf2e-dorako-ui", "theme.app-theme") : false;
+  if (appTheme === "foundry2") {
     background.lineStyle((1 * gridScale) / 2, 0x302831, 1, 0);
     background.drawCircle(effectIcon.position.x, effectIcon.position.y, r + 1 * gridScale);
     background.beginFill(0x0b0a13);
     background.drawCircle(effectIcon.position.x, effectIcon.position.y, r + 1 * gridScale);
     background.endFill();
     return;
-  } else if (theme === "light-theme" || theme == "dark-theme") {
+  } else if (appTheme === "crb") {
     background.lineStyle((1 * gridScale) / 2, 0x956d58, 1, 1);
     background.drawCircle(effectIcon.position.x, effectIcon.position.y, r + 1 * gridScale);
     background.lineStyle((1 * gridScale) / 2, 0xe9d7a1, 1, 0);
     background.drawCircle(effectIcon.position.x, effectIcon.position.y, r + 1 * gridScale);
     background.beginFill(0x956d58);
+    background.drawCircle(effectIcon.position.x, effectIcon.position.y, r + 1 * gridScale);
+    background.endFill();
+    return;
+  } else if (appTheme === "bg3") {
+    background.lineStyle((1 * gridScale) / 2, 0xb68258, 1, 0);
+    background.drawCircle(effectIcon.position.x, effectIcon.position.y, r + 1 * gridScale);
+    background.beginFill(0x000000);
     background.drawCircle(effectIcon.position.x, effectIcon.position.y, r + 1 * gridScale);
     background.endFill();
     return;