+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.1.3
+
+- (Fix) Fixed an issue where the .dorako-ux (for restructuring NPC sheets) class would be added to PC sheets.
+- (New) Added .blind and .whisper classes to the rolltype indicators, so they can be themed.
+
 # 1.1.2
 
 - (Fix) Migrated some styling from Dorako UI, to ensure integrity of the module when used as a stand-alone.